fixed main
