+++ crunch/__init__.py
@@ -1,3 +1,3 @@
-from .__main__ import AUDIT_NAME, OUTPUT_NAME, SONG_SHEET_NAME, main
+from .main import AUDIT_NAME, OUTPUT_NAME, SONG_SHEET_NAME, main
 
 __all__ = ["SONG_SHEET_NAME", "OUTPUT_NAME", "AUDIT_NAME", "main"]
